Calculada luz incidente y arreglado floats en fichero

+++ Trazador.h
@@ -38,5 +38,5 @@ void trazarRayosSombra(Rayo, float, int, Esfera);
 VectorT interseccion(Rayo, Esfera);
 VectorT resolverSegundoGrado(float,float,float);
 void leerFichero();
-void escribirColor(float,float,float, int);
+void escribirColor(int,int,int, int);
 void escribirCabecera();
